hb/pyproject.toml
[tool.poetry]
name = "beetu-v2.0"
version = "0.1.0"
description = "RAG-powered agentic chatbot"
authors = ["Habuild Team"]
readme = "README.md"
license = "MIT"
packages = [{ include = "beetu_v2", from = "src" }]

[tool.poetry.dependencies]
python = "^3.10"
fastapi = "^0.116.1"
uvicorn = "^0.35.0"
openai = "^1.100.2"
langchain = "^0.3.27"
pinecone = ">=7.3.0,<8.0.0"
sqlalchemy = "^2.0.43"
alembic = "^1.16.4"
pgvector = "^0.4.1"
psycopg2-binary = "^2.9.10"
redis = "^6.4.0"
python-dotenv = "^1.1.1"
pydantic-settings = "^2.10.1"
langgraph = "^0.6.6"
langchain-community = "^0.3.27"
python-multipart = "^0.0.20"
langchain-openai = "^0.3.30"
tiktoken = "^0.11.0"
llama-index-core = "^0.13.2"
pypdf = "^6.0.0"
llama-index-readers-file = "^0.5.2"
llama-index-embeddings-openai = "^0.5.0"
llama-index-readers-web = "^0.5.0"
llama-parse = "^0.5.15"
cohere = "^5.17.0"
pypdf2 = "^3.0.1"
matplotlib = "^3.10.6"
pydantic = "^2.11.7"
typing-extensions = "^4.15.0"
google-api-python-client = "^2.181.0"
google-auth-httplib2 = "^0.2.0"
google-auth-oauthlib = "^1.2.2"
pytz = "^2025.2"

[tool.poetry.group.dev.dependencies]
pytest = "^8.4.1"
pytest-asyncio = "^1.1.0"
black = "^25.1.0"
mypy = "^1.17.1"
ruff = "^0.12.9"
pylint = "^3.3.8"
isort = "^6.0.1"

[tool.poetry.scripts]
dev = "beetu_v2.main:main"

[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
